+++ tests/test_mass_cytometry_umap.py
@@ -1,4 +1,3 @@
-
 from click.testing import CliRunner
 
 from mass_cytometry_reduction.cli import main
@@ -8,5 +7,5 @@ def test_main():
     runner = CliRunner()
     result = runner.invoke(main, [])
 
-    assert result.output == '()\n'
+    assert result.output == "()\n"
     assert result.exit_code == 0