+++ CHANGELOG.rst
@@ -2,7 +2,13 @@
 Changelog
 =========
 
+0.1.000 (2021-01-30)
+--------------------
+
+* First version capable of ingesting data, performing dimensional reduction,
+  and writing results to file.
+
 0.0.999 (2021-01-29)
 --------------------
 
-* First release on PyPI.
+* First commit.